Add professional WPF UI with Material Design styling and icon generation
Major UI overhaul replacing basic TaskDialog popups with modern WPF dialogs:

- Add ModernDialog.xaml/xaml.cs with Material Design styling
  - Blue gradient headers with shadows and rounded corners
  - Status cards, info sections, and stats grids
  - Professional button styling with hover effects

- Add IconGenerator.cs for programmatic vector icon creation
  - Connect icon (green play button)
  - Disconnect icon (red stop button)
  - Status icon (blue analytics chart)
  - Settings icon (gray gear)
  - Brand icon (3D cube)

- Enhance App.cs with modern ribbon and icon integration
  - Generate and display professional icons in ribbon
  - Add tooltips and descriptions
  - Create Connection and Tools panels

- Update BridgeCommands.cs with modern dialog integration
  - Replace all TaskDialog calls with ModernDialog
  - Add status cards with color-coded icons
  - Professional success/error messaging

- Add AdditionalCommands.cs for Settings, Help, and About dialogs
  - Settings dialog with server configuration
  - Help dialog with documentation and command overview
  - About dialog with version info and credits

- Configure RevitBridge.csproj for WPF support
  - Enable UseWPF for XAML compilation
  - Add PresentationCore, PresentationFramework, WindowsBase references
  - Temporarily exclude Phase 1-4 commands (API compatibility issues)

- Fix namespace conflicts
  - Add MediaColor/MediaBrush aliases to resolve Revit.DB.Color vs WPF Color conflicts

Build: 0 errors, 82 warnings (nullable - acceptable)
Status: UI production-ready for testing
